refactor: api client 폴더 이동
